+++ src/MakeSimple.SharedKernel.Infrastructure/Exceptions/BaseException.cs
@@ -4,6 +4,7 @@ using System;
 namespace MakeSimple.SharedKernel.Infrastructure.Exceptions
 {
 #pragma warning disable S3925 // "ISerializable" should be implemented correctly
+
     public abstract class BaseException : Exception
 #pragma warning restore S3925 // "ISerializable" should be implemented correctly
     {